pick node nearest to goal if no path to goal could be found
This change is needed, if the robot navigates in unknown terrain
and the path to the goal has not yet been discovered. This change
should led to an behaviour, were the robot approaches unknown
terrain and stop right in front of it, if it can not gather any
new information about it.