implemented reuse of Treenodes
In case that an expanded subtree is found that has worse
cost than the current expanded node, this subtree will
be dropped. Until now the nodes of the subtree were not
reused. This commit changes this behaviour.
In my test the reuse of nodes showed a memory footprint
reduction of ~30%