Allow projection function to return multiple poses
This fixes a design flaw. In the case that a robot
can drive backwards, it may either turn around and
drive into the given direction, or just drive backwards.
The first case results in a different orientation than the
second case. As in the projection function there is not
enought information available to discard one pojection
candidate it must return both.